refactor(iota-json-rpc): rename coin_api module to utils
